Rename test file and get rid of old one
